+++ anybox/recipe/openerp/testing.py
@@ -1,5 +1,9 @@
 """Utilities for unit tests."""
 import os
+import unittest
+import sys
+import shutil
+from tempfile import mkdtemp
 from anybox.recipe.openerp import vcs
 
 class FakeRepo(vcs.BaseRepo):
@@ -10,7 +14,10 @@ class FakeRepo(vcs.BaseRepo):
 
     vcs_control_dir = '.fake'
 
+    revision = 'fakerev'
+
     def get_update(self, revision):
+        self.revision = revision
         if not os.path.isdir(self.target_dir):
             os.mkdir(self.target_dir)
         control = os.path.join(self.target_dir, self.vcs_control_dir)
@@ -23,6 +30,9 @@ class FakeRepo(vcs.BaseRepo):
             options['clear_locks'] = self.clear_locks
         self.log.append((self.target_dir, self.url, revision, options),)
 
+    def parents(self):
+        return [self.revision]
+
 vcs.SUPPORTED['fakevcs'] = FakeRepo
 
 def get_vcs_log():
@@ -30,4 +40,31 @@ def get_vcs_log():
 
 def clear_vcs_log():
     FakeRepo.log = []
+
+
+class RecipeTestCase(unittest.TestCase):
+    """A base setup for tests of recipe classes"""
+
+    def setUp(self):
+        b_dir = self.buildout_dir = mkdtemp('test_oerp_base_recipe')
+        develop_dir = os.path.join(b_dir, 'develop-eggs')
+        os.mkdir(develop_dir)
+        self.buildout = {}
+        self.buildout['buildout'] = {
+            'directory': b_dir,
+            'offline': False,
+            'parts-directory': os.path.join(b_dir, 'parts'),
+            'bin-directory': os.path.join(b_dir, 'bin'),
+            'find-links': '',
+            'allow-hosts': '',
+            'eggs-directory': 'eggs',
+            'develop-eggs-directory': develop_dir,
+            'python': 'main_python',
+            }
+
+        self.buildout['main_python'] = dict(executable=sys.executable)
+
+    def tearDown(self):
+        clear_vcs_log()
+        shutil.rmtree(self.buildout_dir)