+++ anybox/recipe/openerp/tests/test_base.py
@@ -1,11 +1,12 @@
 import unittest
 
 import os
-import shutil
-from tempfile import mkdtemp
+import subprocess
+from ConfigParser import ConfigParser, NoOptionError
 from anybox.recipe.openerp.server import BaseRecipe
 from anybox.recipe.openerp.base import main_software
-
+from anybox.recipe.openerp.base import GP_VCS_EXTEND_DEVELOP
+from anybox.recipe.openerp.testing import RecipeTestCase
 
 class TestingRecipe(BaseRecipe):
     """A subclass with just enough few defaults for unit testing."""
@@ -13,20 +14,7 @@ class TestingRecipe(BaseRecipe):
     archive_filenames = {'6.1': 'blob-%s.tgz'}
     archive_nightly_filenames = {'6.1': '6-1-nightly-%s.tbz'}
 
-class TestBaseRecipe(unittest.TestCase):
-
-    def setUp(self):
-        b_dir = self.buildout_dir = mkdtemp('test_oerp_base_recipe')
-        self.buildout = {}
-        self.buildout['buildout'] = {
-            'directory': b_dir,
-            'offline': False,
-            'parts-directory': os.path.join(b_dir, 'parts'),
-            'bin-directory': os.path.join(b_dir, 'bin'),
-            }
-
-    def tearDown(self):
-        shutil.rmtree(self.buildout_dir)
+class TestBaseRecipe(RecipeTestCase):
 
     def make_recipe(self, name='openerp', **options):
         self.recipe = TestingRecipe(self.buildout, name, options)
@@ -102,7 +90,6 @@ class TestBaseRecipe(unittest.TestCase):
         self.assertEquals(bcn(('--config=6.1.cfg',)), '6.1.cfg')
         self.assertEquals(bcn(('--config=6.1.cfg', '-o')), '6.1.cfg')
 
-
     def test_parse_addons_revisions(self):
         """Test both parse_addons and parse_revisions."""
         self.make_recipe(version='bzr lp:openobject-server server last:1')
@@ -122,3 +109,119 @@ class TestBaseRecipe(unittest.TestCase):
         self.assertEquals(recipe.sources['addons-specific'],
                           ('hg', ('http://some/repo', '1.0'), {'opt': 'spam'}))
         self.assertEquals(recipe.sources[main_software][1][1], '1111')
+
+    def test_freeze_egg_versions(self):
+        """Test that an egg requirement is properly dumped with its version.
+
+        Since nose is our test launcher, we use it as an example, because it
+        must be available without downloading anything."""
+        conf = ConfigParser()
+        conf.add_section('freeze')
+        self.make_recipe(version='6.1-1')
+        self.recipe.options['eggs'] = 'nose'
+        self.recipe.install_requirements() # to get 'ws' attribute
+        self.recipe._freeze_egg_versions(conf, 'freeze')
+        try:
+            nose_version = conf.get('freeze', 'nose')
+        except NoOptionError:
+            self.fail("Expected version of nose egg not dumped !")
+        import nose
+        # GR: maybe that'll turn out to be frail
+        self.assertEquals(nose_version, nose.__version__)
+
+    def test_freeze_egg_versions_merge(self):
+        """Test that freezing of egg versions keeps eggs already dumped.
+
+        very much similar to test_freeze_egg_versions.
+        """
+        conf = ConfigParser()
+        conf.add_section('freeze')
+        conf.set('freeze', 'some.distribution', '1.0alpha')
+        self.make_recipe(version='6.1-1')
+        self.recipe.options['eggs'] = 'nose'
+        self.recipe.install_requirements() # to get 'ws' attribute
+        self.recipe._freeze_egg_versions(conf, 'freeze')
+        try:
+            version = conf.get('freeze', 'some.distribution')
+        except NoOptionError:
+            self.fail("Expected version of 'some.distribution' not kept !")
+        self.assertEquals(version, '1.0alpha')
+
+    def test_freeze_vcs_source(self):
+        self.make_recipe(version='6.1-1')
+        b_dir = self.recipe.buildout_dir
+        repo_path = os.path.join(b_dir, 'custom')
+        subprocess.check_call(['hg', 'init', repo_path])
+        with open(os.path.join(repo_path, 'somefile'), 'w') as f:
+            f.write('content')
+        subprocess.check_call(['hg', '--cwd', repo_path,
+                               'commit', '-A', '-m', 'somerev'])
+
+        rev = self.recipe._freeze_vcs_source('hg', repo_path)
+        hg = subprocess.Popen(['hg', '--cwd', repo_path, 'diff', '-r', rev],
+                              stdout=subprocess.PIPE,
+                              stderr=subprocess.PIPE)
+        out, err = hg.communicate()
+        if hg.returncode or err:
+            self.fail("Invalid extracted revision: %r" % rev)
+        self.assertEquals(out, '', 'Extracted revision shows some diff')
+
+    def test_freeze_vcs_source_dirty(self):
+        self.make_recipe(version='6.1-1')
+        b_dir = self.recipe.buildout_dir
+        repo_path = os.path.join(b_dir, 'custom')
+        subprocess.check_call(['hg', 'init', repo_path])
+        with open(os.path.join(repo_path, 'somefile'), 'w') as f:
+            f.write('content')
+        subprocess.check_call(['hg', '--cwd', repo_path,
+                               'commit', '-A', '-m', 'somerev'])
+
+        # modification on tracked file
+        with open(os.path.join(repo_path, 'somefile'), 'w') as f:
+            f.write('changed content')
+        self.assertRaises(RuntimeError,
+                          self.recipe._freeze_vcs_source, 'hg', repo_path)
+        subprocess.check_call(['hg', '--cwd', repo_path, 'revert', '--all'])
+
+        # untracked file
+        with open(os.path.join(repo_path, 'untracked'), 'w') as f:
+            f.write('something else')
+        self.assertRaises(RuntimeError,
+                          self.recipe._freeze_vcs_source, 'hg', repo_path)
+
+    def test_prepare_frozen_buildout(self):
+        self.make_recipe(version='6.1-1')
+        conf = ConfigParser()
+        self.recipe._prepare_frozen_buildout(conf)
+        self.assertTrue('buildout' in conf.sections())
+
+    def test_prepare_frozen_buildout_gp_vcsdevelop(self):
+        self.make_recipe(version='6.1-1')
+        self.recipe.b_options[GP_VCS_EXTEND_DEVELOP] = (""
+            "fakevcs+http://example.com/aeroolib#egg=aeroolib")
+
+        conf = ConfigParser()
+        self.recipe._prepare_frozen_buildout(conf)
+        extends_develop = conf.get('buildout', GP_VCS_EXTEND_DEVELOP)
+        self.assertEquals(extends_develop.strip(),
+                          "fakevcs+http://example.com/aeroolib@fakerev"
+                          "#egg=aeroolib")
+
+    def test_prepare_frozen_buildout_gp_vcsdevelop_already_fixed(self):
+        """Test that prepare_frozen_buildout understands existing pinning.
+
+        One might say that we souldn't touch an existing revision pinning, but
+        a difference can arise from a tag resolution, or simply someone
+        manually updating the repo. In all cases, the instrospected revision
+        will be used.
+        """
+        self.make_recipe(version='6.1-1')
+        self.recipe.b_options[GP_VCS_EXTEND_DEVELOP] = (""
+            "fakevcs+http://example.com/aeroolib@somerev#egg=aeroolib")
+
+        conf = ConfigParser()
+        self.recipe._prepare_frozen_buildout(conf)
+        extends_develop = conf.get('buildout', GP_VCS_EXTEND_DEVELOP)
+        self.assertEquals(extends_develop.strip(),
+                          "fakevcs+http://example.com/aeroolib@fakerev"
+                          "#egg=aeroolib")